feat(exceptions): WIP: create custom ConversionError class
I have created a new custom exception class, `ConversionError`, which will be used to provide semantic, high-level errors for data conversion failures.

This commit includes the initial implementation of the class in the new `system/utils/exceptions.py` module.

The refactoring of the converter functions and their respective tests to use this new exception is still in progress.